+++ src/components/services/ContactsLoader.ts
@@ -102,6 +102,18 @@ const merchantPartners = [
     recordID: ''
   },
   {
+    givenName: 'Fantom/Sonic Bridge',
+    hasThumbnail: true,
+    thumbnailPath: 'fantomsonicbridge.png',
+    emailAddresses: [],
+    postalAddresses: [],
+    middleName: '',
+    company: '',
+    jobTitle: '',
+    familyName: '',
+    recordID: ''
+  },
+  {
     givenName: 'EOS Network',
     hasThumbnail: true,
     thumbnailPath: 'eos-logo-solo-64.png',