+++ src/util/corePlugins.ts
@@ -107,6 +107,7 @@ export const swapPlugins = {
   '0xgasless': ENV['0XGASLESS_INIT'],
 
   cosmosibc: true,
+  fantomsonicbridge: true,
   transfer: true
 }